Fix mypy not running in workflows
There is no step with an ID of "dependencies" so the mypy step was never
running. We could add the ID to the "Install dependencies" step but
that's really not necessary when the job will fail early at that step
anyways if it doesn't succeed.

Since reintroducing mypy here will probably produce a bunch of errors
I'm sticking `continue-on-error: true` on it for now.